move stream_lib and waiter into async_lib
