+++ lychee-lib/src/async_lib/mod.rs
@@ -1,0 +1,6 @@
+//! Library for async utilities which are useful for lychee. These
+//! should only be functions which are missing from other libraries.
+
+pub mod stream;
+pub mod waiter;
+
